chore: extracting navigation from main activity
